Ignore type error for now.
Need to clean up the interface/think of a nice way of doing this.
We need to get the original item back somehow. Previously, this worked
as everything was item based, so we just returned tuples of items.

Now its list based we can't do that. Worse, we don't even get the same
length list back, and its sorted, which makes matching up a bigger pain.

For now, lets ignore it and do a slow but functional way, and figure out
the crash first.