feat(extensions): Add Service_Net + JSON requests (#2045)
For extension management, we need a way to query the `open-vsx.org` directory - this adds a `Service_Net` service project with a `Request.json` function, to help with this. Also adds some test infra and a health-check for making a `json` request.

Unfortunately, there isn't a Luv-based JSON request OCaml module at the moment - we have `fetch`, but it uses `Lwt`, and we need `Lwt.main_run` to execute it - which I'm not sure we can parallelize with our app execution (`Lwt.main_run`) is blocking. So for now - I defer the work of making a request to our bundled `node`. Hopefully, in the future, we can switch to a pure OCaml+Luv-based solution for making requests.