fix(vim/#2162): Crash with arrow keys (#2172)
here were a couple of problems with our existing command-line arrow key handling - the first is that the `cmdpos` wouldn't always be up-to-date, causing us to get a seg fault when expanding for completions. 

Once the seg fault was fixed, another issue was discovered with the arrow keys - the command line had _two_ notions of position - one is the byte position of the cursor relative to the command line text (`cmdpos`), and the other is the position on-screen of the cursor (`cmdspos`). The `cmdspos` isn't really relevant for Onivim, because we have custom rendering outside of a 'terminal grid' - but, the `cmdspos` could stop the cursor moving if `libvim` thought it was out-of-bounds. So the `cmdspos` was also removed.

This brings in https://github.com/onivim/libvim/pull/216, and also adjusts our completion expansion to only trigger at the end of the string, since we don't reliably complete in the middle of a command at the moment.

Fixes #2162