Base file explorer behavior on autoReveal
The `fileExplorerActiveFileSub` is now based on the `autoReveal`
configuration type. If the user supplies true, the file explorer will
highlight and scroll to the file. If the user supplies false, the file
explorer will do nothing and remain static. If the user supplies the
string 'focusNoScroll' then the file explorer will highlight the file
but it will not scroll it into view.