feat(ux): Window-style navigation for external pieces UI (#2436)
More work on the path for #528 - this allows UI components outside the editor text surface (like sidebar panes!) to opt-in to vim-style window navigation. I envision using this to navigate between pseudo-splits in the UI - for example, in search, to jump between the search text box and the results list:

![2020-09-11 09 13 39](https://user-images.githubusercontent.com/13532591/92949104-5ec6e000-f40f-11ea-8db2-69f4e7ac59c8.gif)

This lets features opt-in to handling window navigation in their own way, and then if they can't handle it (ie, it's at the border) - it would bubble back up to be handled by our layout system.

This doesn't implement list navigation yet (ie, the search results aren't actually navigable) - that's the next step.