feat(explorer/#2231): Add explorer.autoReveal configuration (#2414)
* Add autoReveal config option

* Add autoReveal parser

* Reveal focused file based on autoReveal config

If `explorer.autoReveal` is true, the file explorer will scroll and
highlight the file that has focus. If this setting is false, the file
explorer will remain static and not scroll (or currently highlight) the
focused file.

* Update settings markdown document for autoReveal

* Correctly format ternary

* Create `autoReveal` type

Creating an autoReveal type that can be used in the configuration parser
to represent the various settings for autoReveal.

* Implement autoReveal parser

* Update autoReveal configuration docs

* Base file explorer behavior on autoReveal

The `fileExplorerActiveFileSub` is now based on the `autoReveal`
configuration type. If the user supplies true, the file explorer will
highlight and scroll to the file. If the user supplies false, the file
explorer will do nothing and remain static. If the user supplies the
string 'focusNoScroll' then the file explorer will highlight the file
but it will not scroll it into view.

* Add extra verbiage to autoReveal setting documentation

* Remove autoReveal from StoreThread.re

* Move autoReveal setting to File Explorer store

* Update settings.md

* Move setActive into Highlight* autoReveal conditions

* Pull in `esy format` formatting differences

* Add autoReveal parser tests

* Pull in `esy format`

Co-authored-by: Bryan Phelps <bryphe@outrunlabs.com>