refactor(titlebar): Move TitleBar to feature (#2453)
In the move-file-explorer-to-feature work, there was an idiosyncratic dependency on the title store - specifically, triggering an imperative `SetTitle` effect. This decouples that dependency, so that the file explorer and title can be independent of each other.

In addition, this moves the `window.title` configuration setting to the `Feature_TitleBar` feature, migrates the tests to inline tests, and removes the `TitleStoreConnector`.