docs: Fix release steps. (#2500)
Just updates the release step, since the release is now in its own sandbox, you need to run Oni2 in that sandbox once for the binary to exist, to be copied for the release.

Arguably, the scripts should be a little smarter about this (i.e. check, error, say "Run `esy ....`).