refactor(ux): Untangle context keys (#2511)
In wiring up some of the new window navigation / list navigation / tree navigation, the context keys starting to become unruly. We'd often need the `model` to decide which context keys (ie, in the search pane - do we want input context keys or tree context keys), and then, we'd need to use `Schema.map` to supply the model again.

To streamline this, this changes the `contextKeys` function to just return an abstract `WhenExpr.ContextKeys.t`, instead of a polymorphic `WhenExpr.ContextKeys.Schema.t('model)`. This means it's a little bit easier to combine them, especially in the case where we have nested controls.

This was blocking, or at least making much more difficult, the implementation of a search `/` experience in the list / tree views - this refactoring is intended to help smooth the trailing for implementing search in those views.