fix(editor/#2421): Reopen after closing is not working (#2547)
__Issue:__ As described in #2421 - when closing a buffer, and then reopening, the buffer would not open the second time.

__Defect:__ There was an effectful depending on a 'buffer enter' event - this was used when we used the [`msgpack`](https://neovim.io/doc/user/api.html#nvim_buf_attach()) API - we'd get notified that a buffer was loaded via a buffer enter event, and then have a separate update event.

This flow was emulated in `reason-libvim` - having a split between an enter and update event. However, this added some complexity - Onivim relied on a buffer enter event to populate a new buffer, but it was a globably stateful - and sometimes that state wouldn't match up. For #2421 , we wouldn't get that buffer enter event, and an editor wouldn't be opened.

__Fix:__ Remove that stateful buffer-entered event, which involves some refactoring around the pipeline... but it gives us much more flexibility in how we integrate buffers (ie, loading a buffer without an editor).

Fixes #2421