perf(buffer): Remove shaping on all lines (#2563)
While working on porting over the changes in #1969 for word wrap in #1444  , and profiling, I noticed a regression in the benchmarks - the add-large-amounts-of-lines to buffer test was very slow. Additionally, testing with a large file, I noticed hangs on some large files.

I tracked down the cause of the regression to the shaping that happens here: https://github.com/onivim/oni2/blob/21da38a975e0dfa8529619b4aae2b58e446c5d56/src/Core/BufferLine.re#L266

Whenever we load a buffer, or change a line, that shaping code-path is hit - causing us to run the harfbuzz shaper over the entire buffer line. For small files, or small lines, it's fast enough that it slid by unnoticed. However, for large files and large lines, it is very expensive. The `BufferLine` is structured around being incremental - figuring out metrics like character index / byte index mappings, and pixel widths, as needed - so that we can avoid or defer some of the computation. The shaping, though, is always run for the entire line.

The shaping is needed so that we can get correct pixel-width calculations for rendered items - however, we can do that on a per UTF-8 character basis (1*), as needed, which means the results are more cache-friendly and can continue to be incremental - we don't need to do all that up-front work. This improves performance for loading large files and lines.

In order to implement this, the `BufferLine` is decoupled from needing to know about indentation or fonts - it really just needs to know the pixel width of characters. This can be abstracted via a `measure` function.

__Caveats:__
- 1: Technically, we'd really want to look at a per-grapheme basis, as some rendered characters are a construction of multiple Unicode-code points (a library like [uuseg](https://github.com/dbuenzli/uuseg) will be helpful here!). This is an enhancement we'll eventually need, but at the moment, graphemes aren't handled in the current code anyway, so this isn't a regression in functionality from what we have. This can be computed without needing to shape the entire text (just the grapheme cluster). This _would_ change the form of the `measure` function - it would need to be able to measure a grapheme cluster (an array of code points), and the buffer line would need to include grapheme detection as iterates through the string.

__TODO:___
- [x] Update test cases
- [x] Can `Oni_Core.Font.t` be removed and replaced with `Service_Font.font`?
- [x] Move measurement cache to `Service_Font.font`/`Oni_Core.Font.t`
  - [x] Add fast-access ASCII cache