feat(diagnostics): Runtime trace logging (#2598)
This adds a switch during runtime to turn on trace logging to a file:

![2020-10-19 12 57 39](https://user-images.githubusercontent.com/13532591/96505437-f8519080-120a-11eb-9908-02c5f956151b.gif)


- Accessible via the `Command+Shift+P`/`Control+Shift+P` command palette - `Debug: Log to trace file`

When this option is selected, an `onivim-trace.log` file will be created in the current working directory, and the trace log level will be enabled. 

The use case for this is to acquire more details around issues like https://github.com/onivim/oni2/issues/2407 - cases where the editor may not have been started with logging, but having the log would be helpful in diagnosing.