refactor(keybindings): Add API to query current bindings & consumed keys (#2853)
This refactors the `EditorInput` and `Feature_Input` modules to expose some additional information, answering the questions:
- What keybindings are currently available, given my context?
- What keys are currently consumed?

This will be useful in a couple ways:
- For the menu work in #2849 - this API can be used to dynamically look up shortcut keys
- For a feature like https://github.com/onivim/oni2/issues/756 , we can show a set of available bindings (just need to decide on the best UX) - useful as a learning feature
- For supercharging sneak (Control+g) - in cases where the gesture would map to a bound command, we could show the shortcut key for it - helping to teach some of the Onivim-specific shortcut keys.

This isn't surfaced, yet, although some of the information is available in the (ugly...) debug input view:
![2020-12-16 15 56 58](https://user-images.githubusercontent.com/13532591/102420651-7df97e80-3fb7-11eb-9758-6623bf162732.gif)

Note that when I press `Control+w`, the list of available bindings is refined to the ones available with a subsequent key press. Likewise, in insert mode, I have a binding for `jj` -> `Escape`. So I think taking this information, augmenting it with some of the built-in Vim keybindings, and bubbling it up to the user in a more useful way could help ease the learning curve