fix(vim/#2968): Cursor management in command-line mode (#3020)
__Issue:__ There were a couple related issues with the cursor in command-line mode:
1) When exiting command-line mode, the cursor could 'flash' at the top of the file
2) When running a search command, the viewport would not scroll to the first match

__Defect:__ Both issues stem from the fact that Onivim was ignoring editor-cursor-position changes during search. 

For the first issue, when exiting command-line mode, there would be a brief transitional state where the editor is focused in CommandLine mode, but no cursor would be available - so it would be rendered at the default (0,0) position. This would be rectified once the mode transition completed and the cursor position is reset. 

For the second issue, `libvim` was properly updating the cursor position when searching, but that cursor position was being ignored.

__Fix:__ Track the editor cursor position during the course of the command-line interaction

Fixes #2968 
Related #1551