refactor(quickmenu): Initial feature implementation (#3228)
There are several long-standing bugs with the quick-open experience that I'd like to address:

- https://github.com/onivim/oni2/issues/3165
- https://github.com/onivim/oni2/issues/2394
- https://github.com/onivim/oni2/issues/2259
- https://github.com/onivim/oni2/issues/1579
- https://github.com/onivim/oni2/issues/1126
https://github.com/onivim/oni2/issues/967

We're missing the ability to have more flexible rendering in the quickmenu items (ie, to show the filename first), to show bound keys, or to show a different set of initial items vs actively filtered items.

This is a refactoring to migrate the quick menu from our old-architecture-style of store connectors, to the newer Feature architecture. The goal with this is to make it easier to fix some of the bugs that have been around for a while, and introduce some improvements to the quick-open and command line user experiences.

This change shouldn't impact functionality, but it migrates both of the snippet menus (insert snippet, configure snippets) to the new architecture.