fix(formatting): Compute minimal updates for buffer updates / shifting (#3260)
In cases where there is a scoped buffer update, like a few lines, Onivim is smart enough to shift 'markers' (syntax highlighting, codelens in #3255 , search highlights).

However, there are some cases where we get a full update - the entire buffer contents is flagged as updated. A couple of these cases are:
- Running `:e!` to re-load the file from disk - this comes through as a refresh of the entire buffer
- Some format providers provide scoped updates (ie, line-by-line). Others, however, like `ocamllsp` just send the whole buffer contents.

The problem with processing the full update is that, today, those don't shift markers - we just clear out syntax highlighting, codelens, etc. This is a poor user experience, because it causes 'flashes' of syntax highlighting disappearing, or codelens disappearing and then re-appearing.

In order to have a better experience - this PR adds the computation of a set of minimal updates from these full updates. After computing minimal updates, Onivim can then properly shift markers, codelens, syntax highlighting, etc.

This is important for #3246 , as this same logic will be used to shift and preserve the cursor position.