perf(buffers): Skip applying full updates if nothing changed
Processing full-buffer updates are expensive - we recompute syntax highlighting for the entire buffer, we recompute minimal updates in the case of formatting, and it requires a full sync with the extension host - all expensive operations!

In some cases, this is necessary - for example, reading new data from the file system might mean the entire buffer has changed. However, in some cases, Onivim is processing full updates when it is not strictly necessary (like 'undo').

This PR does a check to see if the full update actually changes the buffer, and if not, skips applying it.