fix(completion/#3283): Part 2 - Fix completion sort order when running with Emmet (#3298)
__Issue:__ While testing out the configuration recommended for #3283 , I observed some strange ordering with completions.

For example, running `ul>li` gave an unexpected ordering:
![image](https://user-images.githubusercontent.com/13532591/111706102-f530ef80-87fe-11eb-8fa8-32bdcb7c4f5d.png)

In particular, note that just the `li` is highlighted.

This turned out to be a larger issue with how Onivim orders completions. There was an erroneous assumption that, for each provider, the 'meet' could be used to fuzzy-match completions. However, in some cases - like the one above - more of the line needs to be taken into account (up to the insert/replace range). In the `ul>li` case, the snippet provider saw that `>` is not a word token, and picked the meet there - and only used `li` to match against.

The fix here is to take into account, for each completion item, the insertion range, and score against that range.

With that in place, we get a better set of ordering for Emmet:
![image](https://user-images.githubusercontent.com/13532591/111706785-2100a500-8800-11eb-9397-e2db2e830993.png)

And generally better ordering for other providers, like TabNine:

__TabNine - After:__

![image](https://user-images.githubusercontent.com/13532591/111706757-0fb79880-8800-11eb-9130-771134fcc5bd.png)

__TabNine - Before:__

![image](https://user-images.githubusercontent.com/13532591/111706677-eac32580-87ff-11eb-844a-ddaca10a5780.png)