fix(completion/#3136): Case-insensitive completion filtering (#3308)
With the work done in #3298 (and #range , it's now easier to adjust the sort / scoring algorithm to fix issues like #3136 .

This PR implements case-insensitive completion filtering, ie for golang:

![2021-03-19 14 25 01](https://user-images.githubusercontent.com/13532591/111843805-09d4bc80-88bf-11eb-9fa0-b8eeea01c1ee.gif)

and other languages:

![image](https://user-images.githubusercontent.com/13532591/111843767-f9bcdd00-88be-11eb-9c17-6c8d5cb8b2db.png)

The scoring function allows inexact matches, but penalizes the score when the case doesn't match. This gets closer to parity with VSCode's matching, although there are some nice-to-haves that aren't implemented in Onivim yet - inexact matches and misspelling/typo resiliency. Incorporating edit distance into the scoring algorithm could help improve even further, as a next step. 

Fixes #3136