fix(explorer/#3350): Implement Reload command + file system watcher for explorer (#3365)
This implements a manual refresh for the file explorer, as well as an experimental setting for automatically updating the file explorer in response to changes.

To enable the file-explorer auto-update, you can set the following config:

```
"files.useExperimentalFileWatcher": true
```

![2021-04-05 14 29 29](https://user-images.githubusercontent.com/13532591/113634965-9522b180-9624-11eb-8066-f7aea527f4e1.gif)

Hopefully can enable it by default, soon! Keep me posted if you try it out and see any issues.

If that's not set, then there is a new Control+Shift+P command - "Explorer: Refresh". This can be bound to a key with the "workbench.files.action.refreshFilesExplorer", like:

```
{ "key": "R", "command": "workbench.files.action.refreshFilesExplorer", "when": "sideBarFocus && !textInputFocus && activeViewlet == 'workbench.view.explorer'" }
```

Fixes #3350 
Fixes #3253 
Fixes #3158
Fixes #2056
Fixes #792 

__TODO:__
- [x] Set up a proper unique ID for the path subscriptions
- [x] Emit event to trigger the extension file system watcher (so git / SCM extensions know to highlight the new file)
- [x] Add configuration setting for watcher
- [x] Validate reload command