Feature_Search: make Enable Regex button focusable
I also switched out the if statements in the Navigation updater for switch
statements. Not only is this cleaner from a Finite State Machine perspective
IMO, but it also (should be) faster since variants are backed by integers so
it can be a simple jump rather than a conditional jump.