feat(windows): Customizable title bar style (#3765)
__Issue:__ There are several quirks on Windows with the 'undecorated' Windows style - issues like:
- #3730 
- #3071 
- #3063 

__Fix:__ These can be deal-breakers for use as a daily editor, so until we have those fixed for the 'undecorated' Windows, change the default setting to use the native titlebar. 

This introduces a new setting - `window.titleBarStyle` that can be `"native"` or `"custom"`. The `"custom"`  titlebar looks nicer, because it is themed and custom rendered, however it has the above quirks. On Windows, change the default to `"native"`.

In addition, this pushes up the configuration loading sooner in the startup cycle, so we can pick up configuration settings like `"window.titleBarStyle"` prior to opening the window.

__Todo:__
- [x] Fix margin on Windows, based on whether we are using the decorated window or not
- [x] Test OSX
- [x] Test Windows (both settings)
- [x] Test on Linux (both settings)